Add files via upload
I developed this project in 1997. The problem was to transfer the contents of a large amount of Floppy disk from a MC6809 Development System Motorola Exorset 100 compatible (Operating system XDOS)  ----> a IBM old PC working on MSDOS (5.00 ver.)  I developed a special hardware and put on a MSDOS PC  I write here a software in Turbo Pascal 5.0  and on MC6809 development system I Write an other software in Omegasoft Pascal M6809  with a part in Assembler 6809.  My old hardware on PC  now does not work on new PC. So I need to change my old project to mantain the link with the 2 systems and also to permit to download from PC to MC6809 system some code developed on my Laptop.